+++ CHANGELOG.md
@@ -36,6 +36,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed ``sat bootsys boot --stage cabinet-power`` to query CAPMC for power status of
   ComputeModules instead of the unsupported NodeBMC type.
 
+### Fixed
+- Actions which wait on certain conditions (for example within ``sat bootsys``) will
+  fail out more quickly when checking for the completion condition fails
+  irrecoverably, instead of repeating a failing check until waiting times out.
+- The ``ncn-power`` stage of ``sat bootsys`` will now only print one error
+  message if ``impitool`` fails multiple times consecutively for a given
+  component. If the number of failures for a given component exceeds 3, then
+  that component will be marked as failed.
+
 ## [3.10.0] - 2021-09-03
 
 ### Changed
@@ -56,6 +65,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   related to trying to start non-existent Ceph services.
 - Fix a traceback that occurred if the user did not have permission to write to
   the log file.
+- Actions which wait on certain conditions (for example within ``sat bootsys``) will
+  fail out more quickly when checking for the completion condition fails
+  irrecoverably, instead of repeating a failing check until waiting times out.
 
 ### Added
 - Added a check for running SDU sessions to the ``session-checks`` stage of a